% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mar23_pop_df}
\alias{mar23_pop_df}
\title{a snapshot of population data from wikidata}
\format{
data.frame
}
\usage{
mar23_pop_df
}
\description{
a snapshot of population data from wikidata
}
\keyword{datasets}
